+++ boatswain/bcolors.py
@@ -1,6 +1,6 @@
 class bcolors:
-    HEADER = '\033[95m'
-    OKBLUE = '\033[94m'
+    HEADER = '\033[94m'
+    OKBLUE = '\033[34m'
     OKGREEN = '\033[92m'
     WARNING = '\033[93m'
     FAIL = '\033[91m'